feat: add translations for sign up & welcome pages
